Removed all the stuff I don't need for this project.
Added the db Connection using EF.
Created the UserDetail Model

Created the UserDetailsController

Added migration files to allow any user to create the sql local db via the windows power shell in the package manager.

Added views for Index, Create Edit, Details and Delete. ( Planning on keeping only Create, as that is the requirement.)

The uploading of the CSV/excel file is still not working.